obnova standoff 12222121

+++ .config/hypr/scripts/wallpaper.sh
@@ -1,19 +1,26 @@
 #!/bin/bash
 WALLPAPER_DIR="$HOME/wallpaper/"
-#I dont know what the fuck I am doing
+
 menu() {
     find "${WALLPAPER_DIR}" -type f \( -iname "*.jpg" -o -iname "*.jpeg" -o -iname "*.png" -o -iname "*.gif" \) | awk '{print "img:"$0}'
 }
+
 main() {
-    choice=$(menu | wofi -c ~/.config/wofi/wallpaper -s ~/.config/wofi/style-wallpaper.css --show dmenu --prompt "Select Wallpaper:" -n)
+    choice=$(menu | wofi -c ~/.config/wofi/wallpaper -s ~/.config/wofi/style-wallpaper.css --show dmenu --prompt "Select Wallpaper:")
     selected_wallpaper=$(echo "$choice" | sed 's/^img://')
-    swww img "$selected_wallpaper" --transition-type simple
-    wal -i "$selected_wallpaper"
-    pkill swaync
-    swaync
-    pkill waybar
-    waybar
-    walogram
+    
+    if [[ -n "$selected_wallpaper" && -f "$selected_wallpaper" ]]; then
+        swww img "$selected_wallpaper" --transition-type simple
+        wal -i "$selected_wallpaper"
+        walogram
+        pkill swaync && swaync
+        pkill waybar 
+        waybar
+        pkill hypr-dock
+        hypr-dock
+        
+    fi
 }
+
 main
